+++ plugins/action/sda_fabric_zones_info.py
@@ -27,8 +27,8 @@ argument_spec = dnac_argument_spec()
 argument_spec.update(dict(
     id=dict(type="str"),
     siteId=dict(type="str"),
-    offset=dict(type="int"),
-    limit=dict(type="int"),
+    offset=dict(type="float"),
+    limit=dict(type="float"),
     headers=dict(type="dict"),
 ))
